Split GetPerformanceReport into several routines
