+++ spec/cassettes/Tasks_GetCcPerformanceReport/with_book.yml
